Add native handle.
